requirements.txt
# ==========================================
# 📦 Bugra-Bot Dependencies
# ==========================================

# Core
ccxt>=4.0.0
pandas>=2.0.0
pandas_ta>=0.3.14b
numpy>=1.24.0

# HTTP / Async
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Backtest (optional)
# matplotlib>=3.8.0
